reorganizing root directory
