--- README.md
@@ -1,94 +1,0 @@
-# Mathematical Research Workspace
-## Millennium Prize Problems & TiddlyWiki Evolution
-
-This workspace contains revolutionary mathematical research and supporting infrastructure for solving the Millennium Prize Problems through innovative computational and publication methods.
-
----
-
-## Project Organization
-
-### 🧮 [millennium-mathematics/](millennium-mathematics/)
-**PRIMARY PROJECT** - Millennium Prize Problem solutions and mathematical publication system
-
-- **Mathematical Proofs** - Revolutionary solutions to unsolved mathematical problems
-- **Symbolic Calculus Engine** - Custom differential calculus implementation  
-- **TiddlyWiki Publication** - Professional mathematical knowledge presentation
-- **Unit Test Verification** - Automated theorem validation
-- **Clay Institute Submission** - Formal mathematical publication pipeline
-
-### 🔬 [tiddlywiki-evolution-research/](tiddlywiki-evolution-research/)
-**SUPPORTING INFRASTRUCTURE** - Advanced TiddlyWiki architectural research
-
-- **Performance Optimizations** - Lazy-loading, memory efficiency (✅ Implemented)
-- **Character Encoding Solutions** - Filesystem-safe mathematical symbols (✅ Implemented)
-- **Population-Based Collaboration** - Mathematical community validation systems
-- **Semantic Git Vectors** - Version control for mathematical knowledge evolution
-- **Mathematical Infrastructure** - Foundation for publication and collaboration systems
-
----
-
-## Core Innovation
-
-### Mathematical Breakthrough
-Using **symbolic differential calculus implemented in code** with **unit test verification** to prove fundamental mathematical theorems, combined with **TiddlyWiki's interconnected knowledge structure** for presenting complex mathematical arguments.
-
-### Publication Revolution
-Transforming mathematical publication from static linear papers to **interactive, interconnected, explorable mathematical knowledge networks** that enable:
-
-- **Non-linear proof exploration**
-- **Automated theorem verification**  
-- **Community-based peer review**
-- **Cross-referenced mathematical concepts**
-- **Population-based validation**
-
----
-
-## Research Philosophy
-
-*"When you trust in math, the proof is in the pudding."*
-
-This work represents a fundamental paradigm shift in mathematical research methodology:
-
-1. **Computational Verification** - Every theorem backed by code
-2. **Collaborative Knowledge Networks** - TiddlyWiki's interconnected structure  
-3. **Population-Based Validation** - Community mathematical consensus
-4. **Open Source Mathematics** - Transparent, verifiable mathematical progress
-
----
-
-## Current Status
-
-🎯 **ACTIVE RESEARCH** - Mathematical breakthroughs in development  
-🧮 **SYMBOLIC CALCULUS** - Core computational engine under construction  
-📝 **TIDDLYWIKI INTEGRATION** - Mathematical publication system being designed  
-🔬 **INFRASTRUCTURE COMPLETE** - Supporting research provides solid foundation  
-🏆 **CLAY SUBMISSION** - Targeting professional Millennium Prize Problem solutions
-
----
-
-## Getting Started
-
-### For Mathematical Work
-```bash
-cd millennium-mathematics/
-npm install
-node setup-mathematical-wiki.js
-```
-
-### For Infrastructure Research  
-```bash
-cd tiddlywiki-evolution-research/
-# Review research reports and implementations
-```
-
-### For TiddlyWiki Core Development
-```bash
-# This workspace includes complete TiddlyWiki source with research modifications
-npm test  # Verify all 1,360 core tests still pass
-```
-
----
-
-**Mission:** Advance human mathematical understanding while revolutionizing how mathematical knowledge is created, verified, and shared.
-
-**Research continues...**
