+++ docs/tidtweaks/recursive-palette-encoder.js
