+++ docs/tidtweaks/rosetta-encoder-v2.js
