+++ docs/tidtweaks/rosetta-encoder.js
