+++ docs/tidtweaks/semantic-git-vectors-research-report.md
