--- existential_math/__init__.py
@@ -1,37 +1,0 @@
-#!/usr/bin/env python3
-"""
-Existential Mathematics Package
-==============================
-
-A revolutionary framework for targeting the Riemann Hypothesis
-through evolutionary attractor dynamics from the ground state ∃.
-
-This package implements:
-- Existential core engine (∃ as ground state)
-- Zeta function attractor analysis
-- Noether symmetry conservation laws
-- Unified RH proof via evolutionary dynamics
-"""
-
-from .core import (
-    ExistentialCore,
-    SymbolicEvolutionEngine,
-    MathematicalVacuum,
-    modular_transformation,
-    critical_line_symmetry,
-    zeta_evolution_operator
-)
-
-from .zeta_analysis import (
-    ZetaAttractorEngine,
-    run_zeta_analysis
-)
-
-from .noether import (
-    NoetherSymmetryEngine,
-    run_noether_analysis
-)
-
-__version__ = "1.0.0"
-__author__ = "Existential Mathematics Team"
-__description__ = "Revolutionary framework for RH via evolutionary dynamics"
