--- existential_math/core.py
@@ -1,227 +1,0 @@
-#!/usr/bin/env python3
-"""
-Existential Mathematics Core Engine
-==================================
-
-The ground state evolution engine where ∃ (there exists) is the vacuum state
-and all mathematics emerges through evolutionary attractors.
-
-This is the foundation for targeting the Riemann Hypothesis through
-symbolic evolution and attractor dynamics.
-"""
-
-from sympy import symbols, simplify, diff, exp, I, pi, sin, cos, log, zeta, re, im
-from sympy.abc import x, y, z, s, t
-from collections import defaultdict
-
-class ExistentialCore:
-    """The ground state evolution engine from ∃"""
-    
-    def __init__(self):
-        self.ground_state = "∃"  # Vacuum state - "there exists"
-        self.evolution_history = [self.ground_state]
-        self.attractors = []
-        self.symmetries = []
-        self.conserved_quantities = []
-        
-    def evolve_from_existence(self, target, description=""):
-        """Evolve from ∃ to a mathematical concept"""
-        current = self.evolution_history[-1]
-        evolution_step = f"{current} → {target}"
-        
-        self.evolution_history.append(target)
-        
-        return {
-            'from': current,
-            'to': target,
-            'evolution': evolution_step,
-            'description': description
-        }
-    
-    def is_attractor(self, expression):
-        """Check if expression is a mathematical attractor (fixed point)"""
-        try:
-            simplified = simplify(expression)
-            return simplified == expression
-        except:
-            return False
-    
-    def track_symmetry(self, expression, symmetry_type):
-        """Track symmetry preservation under evolution"""
-        self.symmetries.append({
-            'expression': str(expression),
-            'symmetry': symmetry_type,
-            'preserved': True
-        })
-    
-    def add_conserved_quantity(self, quantity, description):
-        """Add a conserved quantity (Noether theorem)"""
-        self.conserved_quantities.append({
-            'quantity': str(quantity),
-            'description': description
-        })
-
-class SymbolicEvolutionEngine:
-    """Engine for symbolic evolution with attractor dynamics"""
-    
-    def __init__(self):
-        self.existential_core = ExistentialCore()
-        self.evolution_operators = []
-        self.metric_tensor = None
-        self.curvature_trace = []
-        
-    def add_evolution_operator(self, operator_func, name):
-        """Add an evolution operator (like F_{2k} or PSL(2,ℤ))"""
-        self.evolution_operators.append({
-            'function': operator_func,
-            'name': name
-        })
-    
-    def evolve_symbolic_expression(self, expression, max_steps=10):
-        """Evolve a symbolic expression through attractor dynamics"""
-        evolution_trace = [expression]
-        
-        for step in range(max_steps):
-            current = evolution_trace[-1]
-            
-            # Apply evolution operators
-            for operator in self.evolution_operators:
-                try:
-                    evolved = operator['function'](current)
-                    evolution_trace.append(evolved)
-                    
-                    # Check if we've reached an attractor
-                    if self.existential_core.is_attractor(evolved):
-                        return {
-                            'evolution_trace': evolution_trace,
-                            'attractor': evolved,
-                            'steps_to_attractor': step + 1,
-                            'operator': operator['name']
-                        }
-                except:
-                    continue
-        
-        return {
-            'evolution_trace': evolution_trace,
-            'attractor': None,
-            'steps_to_attractor': None,
-            'operator': None
-        }
-    
-    def compute_symbolic_curvature(self, expression):
-        """Compute symbolic curvature using Ricci-like tensor"""
-        try:
-            # Simplified curvature: trace of second derivatives
-            vars = list(expression.free_symbols)
-            if len(vars) >= 2:
-                curvature = 0
-                for var in vars[:2]:
-                    curvature += diff(diff(expression, var), var)
-                return simplify(curvature) if curvature != 0 else 0
-        except:
-            return 0
-    
-    def ricci_flow_step(self, expression, learning_rate=0.01):
-        """Apply one step of symbolic Ricci flow"""
-        curvature = self.compute_symbolic_curvature(expression)
-        
-        # Ricci flow: ∂g/∂t = -2Ric
-        # In symbolic space: evolve toward minimal curvature
-        if curvature != 0:
-            # Move toward zero curvature
-            try:
-                evolved = expression - learning_rate * curvature
-                return simplify(evolved) if evolved != 0 else expression
-            except:
-                return expression
-        else:
-            return expression  # Already at minimal curvature
-    
-    def find_critical_line_symmetry(self, expression):
-        """Find symmetry with respect to critical line Re(s) = 1/2"""
-        try:
-            # For zeta function: check if ζ(s) = ζ(1-s) on critical line
-            if hasattr(expression, 'free_symbols') and 's' in expression.free_symbols:
-                # Check symmetry: s → 1-s
-                symmetric = expression.subs(s, 1-s)
-                return simplify(expression) == simplify(symmetric)
-        except:
-            return False
-        
-        return False
-
-class MathematicalVacuum:
-    """The vacuum state of mathematics - the ground state from which all emerges"""
-    
-    def __init__(self):
-        self.core = ExistentialCore()
-        self.evolution_engine = SymbolicEvolutionEngine()
-        
-    def seed_from_existence(self, concept, description=""):
-        """Seed a mathematical concept from ∃"""
-        evolution = self.core.evolve_from_existence(concept, description)
-        return evolution
-    
-    def evolve_mathematical_structure(self, structure_name, evolution_path):
-        """Evolve a mathematical structure through its path"""
-        print(f"Evolving {structure_name} from ∃:")
-        
-        for step in evolution_path:
-            evolution = self.core.evolve_from_existence(step, f"Evolution of {structure_name}")
-            print(f"  {evolution['evolution']}")
-        
-        return evolution_path
-    
-    def find_universal_attractors(self):
-        """Find universal mathematical attractors"""
-        universal_attractors = [
-            (1, "Unity"),
-            (0, "Zero"),
-            (pi, "Pi"),
-            (exp(1), "Euler's number"),
-            (I, "Imaginary unit"),
-            (exp(I*pi) + 1, "Euler's identity")
-        ]
-        
-        found_attractors = []
-        for expr, name in universal_attractors:
-            if self.core.is_attractor(expr):
-                found_attractors.append({
-                    'name': name,
-                    'expression': str(expr),
-                    'type': 'universal_attractor'
-                })
-        
-        return found_attractors
-
-# Evolution operators for targeting Riemann Hypothesis
-def modular_transformation(expression):
-    """PSL(2,ℤ) modular transformation: s → (as + b)/(cs + d)"""
-    try:
-        a, b, c, d = symbols('a b c d')
-        # Apply modular transformation
-        transformed = expression.subs(s, (a*s + b)/(c*s + d))
-        return simplify(transformed)
-    except:
-        return expression
-
-def critical_line_symmetry(expression):
-    """Symmetry with respect to critical line Re(s) = 1/2"""
-    try:
-        # Check if expression is symmetric under s → 1-s
-        symmetric = expression.subs(s, 1-s)
-        return simplify(expression) == simplify(symmetric)
-    except:
-        return False
-
-def zeta_evolution_operator(expression):
-    """Evolution operator specifically for zeta function"""
-    try:
-        # Apply functional equation: ζ(s) = 2^s π^(s-1) sin(πs/2) Γ(1-s) ζ(1-s)
-        # This is the key symmetry for Riemann Hypothesis
-        if 'zeta' in str(expression):
-            # Simplified evolution: ζ(s) → ζ(s+1)
-            evolved = expression.subs(s, s + 1)
-            return simplify(evolved)
-    except:
-        return expression
