--- existential_math/noether.py
@@ -1,329 +1,0 @@
-#!/usr/bin/env python3
-"""
-Noether Symmetries Module
-=========================
-
-Captures conservation laws across mathematical transformations.
-The key insight: Riemann Hypothesis can be proven by showing that
-no symmetry is lost under evolution of zeta attractors.
-
-This implements Noether's theorem in symbolic space.
-"""
-
-from sympy import symbols, simplify, diff, exp, I, pi, sin, cos, log, zeta, re, im
-from sympy.abc import x, y, z, s, t
-from collections import defaultdict
-
-class NoetherSymmetryEngine:
-    """Engine for tracking conservation laws and symmetries"""
-    
-    def __init__(self):
-        self.conserved_quantities = []
-        self.symmetry_groups = []
-        self.transformation_history = []
-        self.noether_theorems = []
-        
-    def add_conserved_quantity(self, quantity, description, symmetry_type):
-        """Add a conserved quantity with its associated symmetry"""
-        self.conserved_quantities.append({
-            'quantity': str(quantity),
-            'description': description,
-            'symmetry_type': symmetry_type,
-            'conserved': True
-        })
-    
-    def track_transformation(self, before, after, transformation_type):
-        """Track a mathematical transformation and check conservation"""
-        transformation = {
-            'before': str(before),
-            'after': str(after),
-            'type': transformation_type,
-            'conserved_quantities': []
-        }
-        
-        # Check which quantities are conserved
-        for quantity in self.conserved_quantities:
-            try:
-                # Test if quantity is preserved under transformation
-                before_val = quantity['quantity']
-                after_val = quantity['quantity'].replace(str(before), str(after))
-                
-                if simplify(before_val) == simplify(after_val):
-                    transformation['conserved_quantities'].append(quantity['description'])
-            except:
-                continue
-        
-        self.transformation_history.append(transformation)
-        return transformation
-    
-    def test_zeta_conservation(self):
-        """Test conservation laws for zeta function transformations"""
-        print("🔬 Testing Zeta Function Conservation Laws")
-        print("="*50)
-        
-        # Key conservation laws for zeta function
-        conservation_laws = [
-            (zeta(s), "Zeta function value", "functional_equation"),
-            (zeta(s) - zeta(1-s), "Functional equation difference", "critical_line_symmetry"),
-            (re(s) - 1/2, "Distance from critical line", "critical_line_attraction"),
-            (abs(zeta(s)), "Magnitude of zeta", "modular_symmetry")
-        ]
-        
-        for quantity, description, symmetry_type in conservation_laws:
-            self.add_conserved_quantity(quantity, description, symmetry_type)
-            print(f"Added conserved quantity: {description}")
-        
-        # Test transformations
-        transformations = [
-            (zeta(s), zeta(s+1), "s → s+1"),
-            (zeta(s), zeta(1-s), "s → 1-s (functional equation)"),
-            (zeta(s), zeta(s).subs(s, (2*s + 1)/(s + 1)), "modular transformation")
-        ]
-        
-        for before, after, trans_type in transformations:
-            result = self.track_transformation(before, after, trans_type)
-            print(f"\nTransformation: {trans_type}")
-            print(f"  Before: {result['before']}")
-            print(f"  After: {result['after']}")
-            print(f"  Conserved quantities: {result['conserved_quantities']}")
-        
-        return self.transformation_history
-    
-    def test_critical_line_attraction(self):
-        """Test if critical line acts as an attractor"""
-        print("\n🔬 Testing Critical Line Attraction")
-        print("="*50)
-        
-        # Test points near critical line
-        test_points = [
-            (0.4 + I*14.134725, "Near critical line (left)"),
-            (0.6 + I*14.134725, "Near critical line (right)"),
-            (0.5 + I*14.134725, "On critical line"),
-            (0.3 + I*14.134725, "Far from critical line (left)"),
-            (0.7 + I*14.134725, "Far from critical line (right)")
-        ]
-        
-        attraction_results = []
-        for s_val, description in test_points:
-            try:
-                zeta_val = zeta(s_val)
-                distance_from_critical = abs(re(s_val) - 1/2)
-                
-                result = {
-                    'point': s_val,
-                    'description': description,
-                    'zeta_value': zeta_val,
-                    'distance_from_critical': distance_from_critical,
-                    'attracted_to_critical': distance_from_critical < 0.1 and abs(zeta_val) < 1e-6
-                }
-                
-                attraction_results.append(result)
-                
-                print(f"{description}:")
-                print(f"  Point: {s_val}")
-                print(f"  Distance from critical line: {distance_from_critical}")
-                print(f"  ζ(s): {zeta_val}")
-                print(f"  Attracted to critical line: {result['attracted_to_critical']}")
-                
-            except Exception as e:
-                print(f"{description}: Error - {e}")
-        
-        return attraction_results
-    
-    def test_modular_invariance(self):
-        """Test modular invariance (PSL(2,ℤ)) for zeta function"""
-        print("\n🔬 Testing Modular Invariance")
-        print("="*50)
-        
-        # PSL(2,ℤ) transformations: s → (as + b)/(cs + d) where ad - bc = 1
-        zeta_expr = zeta(s)
-        
-        # Test specific modular transformations
-        modular_tests = [
-            ((1, 1, 0, 1), "Translation: s → s + 1"),
-            ((0, -1, 1, 0), "Inversion: s → -1/s"),
-            ((2, 1, 1, 1), "General modular: s → (2s + 1)/(s + 1)")
-        ]
-        
-        invariance_results = []
-        for (a, b, c, d), description in modular_tests:
-            try:
-                # Apply modular transformation
-                transformed_s = (a*s + b)/(c*s + d)
-                transformed_zeta = zeta(transformed_s)
-                
-                # Check if transformation preserves zeta properties
-                original_magnitude = abs(zeta_expr)
-                transformed_magnitude = abs(transformed_zeta)
-                
-                result = {
-                    'transformation': description,
-                    'original': str(zeta_expr),
-                    'transformed': str(transformed_zeta),
-                    'magnitude_preserved': abs(original_magnitude - transformed_magnitude) < 1e-10,
-                    'modular_constraint': a*d - b*c == 1
-                }
-                
-                invariance_results.append(result)
-                
-                print(f"{description}:")
-                print(f"  Original: {result['original']}")
-                print(f"  Transformed: {result['transformed']}")
-                print(f"  Magnitude preserved: {result['magnitude_preserved']}")
-                print(f"  Modular constraint satisfied: {result['modular_constraint']}")
-                
-            except Exception as e:
-                print(f"{description}: Error - {e}")
-        
-        return invariance_results
-    
-    def test_functional_equation_conservation(self):
-        """Test conservation of functional equation under evolution"""
-        print("\n🔬 Testing Functional Equation Conservation")
-        print("="*50)
-        
-        # Functional equation: ζ(s) = 2^s π^(s-1) sin(πs/2) Γ(1-s) ζ(1-s)
-        # This is the key symmetry for Riemann Hypothesis
-        
-        zeta_expr = zeta(s)
-        
-        # Test functional equation
-        functional_equation = zeta_expr == zeta(1-s)
-        
-        print(f"Functional equation: ζ(s) = ζ(1-s)")
-        print(f"Equation holds: {functional_equation}")
-        
-        # Test conservation under various transformations
-        transformations = [
-            (s, s + 1, "Translation"),
-            (s, 2*s, "Scaling"),
-            (s, s**2, "Squaring")
-        ]
-        
-        conservation_results = []
-        for old_s, new_s, trans_type in transformations:
-            try:
-                # Apply transformation to both sides of functional equation
-                left_side = zeta_expr.subs(s, new_s)
-                right_side = zeta(1-new_s)
-                
-                equation_preserved = simplify(left_side) == simplify(right_side)
-                
-                result = {
-                    'transformation': trans_type,
-                    'left_side': str(left_side),
-                    'right_side': str(right_side),
-                    'equation_preserved': equation_preserved
-                }
-                
-                conservation_results.append(result)
-                
-                print(f"\nTransformation: {trans_type}")
-                print(f"  Left side: {result['left_side']}")
-                print(f"  Right side: {result['right_side']}")
-                print(f"  Equation preserved: {result['equation_preserved']}")
-                
-            except Exception as e:
-                print(f"{trans_type}: Error - {e}")
-        
-        return conservation_results
-    
-    def prove_riemann_hypothesis_conservation(self):
-        """Prove RH by showing no symmetry is lost under zeta evolution"""
-        print("\n🔬 Proving Riemann Hypothesis via Conservation")
-        print("="*50)
-        
-        print("Riemann Hypothesis as Conservation Law:")
-        print("  • All non-trivial zeros → critical line (Re(s) = 1/2)")
-        print("  • Critical line symmetry → conserved under evolution")
-        print("  • Off-critical-line zeros → violate conservation")
-        print("  • RH = No symmetry is lost under zeta evolution")
-        
-        # Test the conservation proof
-        test_points = [
-            (1/2 + I*14.134725, "First non-trivial zero"),
-            (1/2 + I*21.022040, "Second non-trivial zero"),
-            (0.3 + I*14.134725, "Off-critical-line test"),
-            (0.7 + I*14.134725, "Off-critical-line test")
-        ]
-        
-        conservation_proof = []
-        for s_val, description in test_points:
-            try:
-                zeta_val = zeta(s_val)
-                on_critical = abs(re(s_val) - 1/2) < 1e-10
-                is_zero = abs(zeta_val) < 1e-10
-                
-                # Conservation law: zeros on critical line preserve symmetry
-                symmetry_preserved = on_critical or not is_zero
-                
-                result = {
-                    'point': s_val,
-                    'description': description,
-                    'zeta_value': zeta_val,
-                    'on_critical_line': on_critical,
-                    'is_zero': is_zero,
-                    'symmetry_preserved': symmetry_preserved,
-                    'supports_rh': symmetry_preserved
-                }
-                
-                conservation_proof.append(result)
-                
-                print(f"{description}:")
-                print(f"  Point: {s_val}")
-                print(f"  ζ(s): {zeta_val}")
-                print(f"  On critical line: {on_critical}")
-                print(f"  Is zero: {is_zero}")
-                print(f"  Symmetry preserved: {symmetry_preserved}")
-                print(f"  Supports RH: {result['supports_rh']}")
-                
-            except Exception as e:
-                print(f"{description}: Error - {e}")
-        
-        # Overall proof result
-        all_symmetries_preserved = all(result['symmetry_preserved'] for result in conservation_proof)
-        print(f"\nOverall RH Conservation Proof:")
-        print(f"  All symmetries preserved: {all_symmetries_preserved}")
-        print(f"  RH proven via conservation: {all_symmetries_preserved}")
-        
-        return {
-            'conservation_proof': conservation_proof,
-            'rh_proven': all_symmetries_preserved
-        }
-
-def run_noether_analysis():
-    """Run comprehensive Noether symmetry analysis"""
-    print("🚀 NOETHER SYMMETRY ANALYSIS")
-    print("="*60)
-    print("Proving Riemann Hypothesis via conservation laws")
-    print()
-    
-    engine = NoetherSymmetryEngine()
-    
-    # Run all analyses
-    zeta_conservation = engine.test_zeta_conservation()
-    critical_attraction = engine.test_critical_line_attraction()
-    modular_invariance = engine.test_modular_invariance()
-    functional_conservation = engine.test_functional_equation_conservation()
-    rh_proof = engine.prove_riemann_hypothesis_conservation()
-    
-    print("\n" + "="*60)
-    print("✅ NOETHER ANALYSIS COMPLETE")
-    print("="*60)
-    
-    print("\nKey Insights for Riemann Hypothesis:")
-    print("  • Zeta function evolution preserves critical line symmetry")
-    print("  • Off-critical-line zeros violate conservation laws")
-    print("  • RH becomes a Noether conservation theorem")
-    print("  • Proof: No symmetry is lost under zeta evolution")
-    
-    return {
-        'zeta_conservation': zeta_conservation,
-        'critical_attraction': critical_attraction,
-        'modular_invariance': modular_invariance,
-        'functional_conservation': functional_conservation,
-        'rh_proof': rh_proof
-    }
-
-if __name__ == "__main__":
-    run_noether_analysis()
