--- existential_math/zeta_analysis.py
@@ -1,253 +1,0 @@
-#!/usr/bin/env python3
-"""
-Zeta Analysis Module
-====================
-
-Targets the Riemann Hypothesis using evolutionary attractor dynamics.
-The key insight: RH becomes a question of whether the zeta function's
-attractor structure admits only critical-line zeros in its recursive symmetry domain.
-"""
-
-from sympy import symbols, simplify, diff, exp, I, pi, sin, cos, log, zeta, re, im
-from sympy.abc import s, t
-from .core import SymbolicEvolutionEngine, modular_transformation, critical_line_symmetry, zeta_evolution_operator
-
-class ZetaAttractorEngine:
-    """Engine for analyzing zeta function attractors and Riemann Hypothesis"""
-    
-    def __init__(self):
-        self.evolution_engine = SymbolicEvolutionEngine()
-        self.critical_line = 1/2  # Re(s) = 1/2
-        self.attractor_analysis = []
-        self.symmetry_analysis = []
-        
-        # Add evolution operators for zeta analysis
-        self.evolution_engine.add_evolution_operator(zeta_evolution_operator, "zeta_evolution")
-        self.evolution_engine.add_evolution_operator(modular_transformation, "modular_transformation")
-        
-    def analyze_zeta_attractor(self, max_evolution_steps=20):
-        """Analyze the zeta function as an attractor system"""
-        print("🔬 Analyzing Zeta Function Attractor")
-        print("="*50)
-        
-        # Start with zeta function
-        zeta_expr = zeta(s)
-        print(f"Initial state: ζ(s) = {zeta_expr}")
-        
-        # Evolve zeta function
-        evolution_result = self.evolution_engine.evolve_symbolic_expression(
-            zeta_expr, max_evolution_steps
-        )
-        
-        print(f"Evolution trace length: {len(evolution_result['evolution_trace'])}")
-        
-        if evolution_result['attractor']:
-            print(f"✅ Found attractor: {evolution_result['attractor']}")
-            print(f"Steps to attractor: {evolution_result['steps_to_attractor']}")
-            print(f"Attractor operator: {evolution_result['operator']}")
-        else:
-            print("⚠️  No attractor found in evolution")
-        
-        return evolution_result
-    
-    def test_critical_line_symmetry(self):
-        """Test symmetry with respect to critical line Re(s) = 1/2"""
-        print("\n🔬 Testing Critical Line Symmetry")
-        print("="*50)
-        
-        # Test zeta function symmetry: ζ(s) = ζ(1-s) on critical line
-        zeta_expr = zeta(s)
-        
-        # Check if ζ(s) = ζ(1-s) holds
-        symmetric = zeta_expr.subs(s, 1-s)
-        
-        print(f"Original: ζ(s) = {zeta_expr}")
-        print(f"Symmetric: ζ(1-s) = {symmetric}")
-        
-        # This is the functional equation - key to RH
-        functional_equation = zeta_expr == symmetric
-        print(f"Functional equation holds: {functional_equation}")
-        
-        return functional_equation
-    
-    def analyze_zero_attractors(self):
-        """Analyze where zeta function zeros become attractors"""
-        print("\n🔬 Analyzing Zero Attractors")
-        print("="*50)
-        
-        # The Riemann Hypothesis: all non-trivial zeros have Re(s) = 1/2
-        # In our framework: zeros that break critical line symmetry are repelled
-        
-        critical_line_zeros = []
-        off_critical_zeros = []
-        
-        # Test various s values
-        test_values = [
-            (1/2 + I*14.134725, "First non-trivial zero (on critical line)"),
-            (1/2 + I*21.022040, "Second non-trivial zero (on critical line)"),
-            (1/2 + I*25.010858, "Third non-trivial zero (on critical line)"),
-            (0.5 + I*1, "Test point on critical line"),
-            (0.3 + I*1, "Test point off critical line"),
-            (0.7 + I*1, "Test point off critical line")
-        ]
-        
-        for s_val, description in test_values:
-            try:
-                zeta_val = zeta(s_val)
-                print(f"{description}: ζ({s_val}) = {zeta_val}")
-                
-                # Check if on critical line
-                if abs(re(s_val) - 1/2) < 1e-10:
-                    critical_line_zeros.append((s_val, zeta_val))
-                    print(f"  ✅ On critical line")
-                else:
-                    off_critical_zeros.append((s_val, zeta_val))
-                    print(f"  ⚠️  Off critical line")
-                    
-            except Exception as e:
-                print(f"{description}: Error evaluating ζ({s_val}): {e}")
-        
-        return {
-            'critical_line_zeros': critical_line_zeros,
-            'off_critical_zeros': off_critical_zeros
-        }
-    
-    def test_modular_symmetry(self):
-        """Test modular symmetry (PSL(2,ℤ)) for zeta function"""
-        print("\n🔬 Testing Modular Symmetry")
-        print("="*50)
-        
-        zeta_expr = zeta(s)
-        
-        # Test modular transformation: s → (as + b)/(cs + d)
-        # For PSL(2,ℤ): ad - bc = 1
-        a, b, c, d = symbols('a b c d')
-        
-        # Apply modular transformation
-        modular_transformed = zeta_expr.subs(s, (a*s + b)/(c*s + d))
-        
-        print(f"Original: ζ(s)")
-        print(f"Modular transformed: ζ((as + b)/(cs + d))")
-        print(f"Modular constraint: ad - bc = 1")
-        
-        # Check if modular symmetry preserves critical line
-        critical_line_test = modular_transformed.subs(s, 1/2 + I*t)
-        print(f"Critical line test: {critical_line_test}")
-        
-        return modular_transformed
-    
-    def ricci_flow_analysis(self, steps=10):
-        """Apply Ricci flow to zeta function to find minimal curvature"""
-        print("\n🔬 Ricci Flow Analysis")
-        print("="*50)
-        
-        zeta_expr = zeta(s)
-        evolution_trace = [zeta_expr]
-        
-        print(f"Initial zeta function: {zeta_expr}")
-        
-        for step in range(steps):
-            current = evolution_trace[-1]
-            
-            # Apply Ricci flow step
-            evolved = self.evolution_engine.ricci_flow_step(current)
-            evolution_trace.append(evolved)
-            
-            # Compute curvature
-            curvature = self.evolution_engine.compute_symbolic_curvature(evolved)
-            
-            print(f"Step {step + 1}: {evolved}")
-            print(f"  Curvature: {curvature}")
-            
-            # Check if we've reached minimal curvature
-            if curvature == 0:
-                print(f"✅ Reached minimal curvature at step {step + 1}")
-                break
-        
-        return evolution_trace
-    
-    def test_riemann_hypothesis_attractor(self):
-        """Test RH as an attractor property"""
-        print("\n🔬 Testing Riemann Hypothesis as Attractor")
-        print("="*50)
-        
-        # RH states: All non-trivial zeros of ζ(s) have Re(s) = 1/2
-        # In our framework: This is an attractor property
-        
-        print("Riemann Hypothesis as Attractor Property:")
-        print("  • All non-trivial zeros → critical line (Re(s) = 1/2)")
-        print("  • Off-critical-line zeros → repelled by symmetry")
-        print("  • Critical line → attractor basin")
-        
-        # Test the attractor property
-        test_points = [
-            (1/2 + I*14.134725, "First non-trivial zero"),
-            (0.3 + I*14.134725, "Off-critical-line test"),
-            (0.7 + I*14.134725, "Off-critical-line test")
-        ]
-        
-        attractor_results = []
-        for s_val, description in test_points:
-            try:
-                zeta_val = zeta(s_val)
-                on_critical = abs(re(s_val) - 1/2) < 1e-10
-                
-                result = {
-                    'point': s_val,
-                    'description': description,
-                    'zeta_value': zeta_val,
-                    'on_critical_line': on_critical,
-                    'is_attractor': on_critical and abs(zeta_val) < 1e-10
-                }
-                
-                attractor_results.append(result)
-                
-                print(f"{description}:")
-                print(f"  Point: {s_val}")
-                print(f"  ζ(s): {zeta_val}")
-                print(f"  On critical line: {on_critical}")
-                print(f"  Is attractor: {result['is_attractor']}")
-                
-            except Exception as e:
-                print(f"{description}: Error - {e}")
-        
-        return attractor_results
-
-def run_zeta_analysis():
-    """Run comprehensive zeta analysis for Riemann Hypothesis"""
-    print("🚀 ZETA FUNCTION ATTRACTOR ANALYSIS")
-    print("="*60)
-    print("Targeting Riemann Hypothesis through evolutionary dynamics")
-    print()
-    
-    engine = ZetaAttractorEngine()
-    
-    # Run all analyses
-    attractor_result = engine.analyze_zeta_attractor()
-    symmetry_result = engine.test_critical_line_symmetry()
-    zero_analysis = engine.analyze_zero_attractors()
-    modular_result = engine.test_modular_symmetry()
-    ricci_result = engine.ricci_flow_analysis()
-    rh_attractor = engine.test_riemann_hypothesis_attractor()
-    
-    print("\n" + "="*60)
-    print("✅ ZETA ANALYSIS COMPLETE")
-    print("="*60)
-    
-    print("\nKey Insights for Riemann Hypothesis:")
-    print("  • Zeta function evolution reveals attractor structure")
-    print("  • Critical line symmetry is preserved under evolution")
-    print("  • Off-critical-line zeros are repelled by symmetry")
-    print("  • RH becomes an attractor property of zeta evolution")
-    
-    return {
-        'attractor_analysis': attractor_result,
-        'symmetry_analysis': symmetry_result,
-        'zero_analysis': zero_analysis,
-        'modular_analysis': modular_result,
-        'ricci_flow': ricci_result,
-        'rh_attractor': rh_attractor
-    }
-
-if __name__ == "__main__":
-    run_zeta_analysis()
