--- existential_math_demo.py
@@ -1,352 +1,0 @@
-#!/usr/bin/env python3
-"""
-Existential Mathematics Demo
-===========================
-
-The revolutionary demonstration of targeting the Riemann Hypothesis
-through evolutionary attractor dynamics from the ground state ∃.
-
-This demo shows how RH becomes a question of whether the zeta function's
-attractor structure admits only critical-line zeros in its recursive symmetry domain.
-"""
-
-import sys
-import os
-sys.path.append(os.path.dirname(os.path.abspath(__file__)))
-
-from existential_math.core import MathematicalVacuum, SymbolicEvolutionEngine
-from existential_math.zeta_analysis import ZetaAttractorEngine, run_zeta_analysis
-from existential_math.noether import NoetherSymmetryEngine, run_noether_analysis
-from sympy import symbols, simplify, diff, exp, I, pi, sin, cos, log, zeta, re, im, Abs
-from sympy.abc import s, t
-import datetime
-
-def print_title(title):
-    """Print title in all caps, no emoji"""
-    import re
-    stripped = re.sub(r'[^A-Za-z]', '', title)
-    print(stripped.upper())
-
-def print_structure(structure):
-    """Print mathematical structure"""
-    import re
-    stripped = re.sub(r'[^-=*_#]', '', structure)
-    print(stripped)
-
-def print_eval(expr, expected, assertion):
-    """Evaluate symbolic expression and print with emoji based on result"""
-    result = simplify(expr)
-    if result == expected:
-        print(f"✅ {assertion}: {result}")
-    else:
-        print(f"❌ {assertion}: {expr} → {result} (expected {expected})")
-
-def assert_existential_foundation():
-    """Assert the existential foundation: ∃ as ground state"""
-    print_title("EXISTENTIAL FOUNDATION")
-    print_structure("="*60)
-    
-    vacuum = MathematicalVacuum()
-    
-    # What is actually mathematically true about the ground state?
-    ground_state = vacuum.core.ground_state
-    print_eval(ground_state, "∃", "Ground state is existential quantifier")
-    
-    # What structures actually emerge from ∃?
-    structures = [
-        ("∃x", "variable_existence"),
-        ("∃x: P(x)", "property_existence"), 
-        ("∃f: f(x) = y", "function_existence"),
-        ("∃ζ: ζ(s)", "zeta_existence")
-    ]
-    
-    for structure, rule in structures:
-        evolution = vacuum.seed_from_existence(structure, "")
-        # What is mathematically true about this evolution?
-        print_eval(evolution['from'], vacuum.core.evolution_history[-2], f"Evolution from {rule}")
-    
-    # What attractors are actually mathematical fixed points?
-    attractors = vacuum.find_universal_attractors()
-    for attractor in attractors:
-        expr = eval(attractor['expression'])
-        print_eval(vacuum.core.is_attractor(expr), True, f"{attractor['name']} is an attractor")
-    
-    return vacuum
-
-def assert_zeta_attractor_properties():
-    """Assert nontrivial properties of zeta function attractors"""
-    print_title("ZETA FUNCTION ATTRACTOR PROPERTIES")
-    print_structure("="*60)
-    
-    engine = ZetaAttractorEngine()
-    
-    # What is mathematically true about zeta function evolution?
-    attractor_result = engine.analyze_zeta_attractor()
-    print_eval(attractor_result['attractor'], None, "Zeta function attractor")
-    
-    # What is mathematically true about critical line symmetry?
-    symmetry_result = engine.test_critical_line_symmetry()
-    print_eval(symmetry_result, True, "Critical line symmetry preserved")
-    
-    # What is mathematically true about zero locations?
-    zero_analysis = engine.analyze_zero_attractors()
-    critical_line_zeros = zero_analysis['critical_line_zeros']
-    print_eval(len(critical_line_zeros), 3, "Number of critical line zeros")
-    
-    # What is mathematically true about modular symmetry?
-    modular_result = engine.test_modular_symmetry()
-    print_eval(len(modular_result), 0, "Number of modular transformations")
-    
-    # What is mathematically true about Ricci flow?
-    ricci_result = engine.ricci_flow_analysis()
-    print_eval(len(ricci_result), 0, "Number of Ricci flow steps")
-    
-    # What is mathematically true about RH as attractor?
-    rh_attractor = engine.test_riemann_hypothesis_attractor()
-    critical_points = [r for r in rh_attractor if r['on_critical_line']]
-    off_critical_points = [r for r in rh_attractor if not r['on_critical_line']]
-    print_eval(len(critical_points) + len(off_critical_points), 0, "Total RH test points")
-    
-    return {
-        'attractor_analysis': attractor_result,
-        'symmetry_analysis': symmetry_result,
-        'zero_analysis': zero_analysis,
-        'modular_analysis': modular_result,
-        'ricci_flow': ricci_result,
-        'rh_attractor': rh_attractor
-    }
-
-def assert_noether_conservation_laws():
-    """Assert nontrivial Noether conservation laws"""
-    print_title("NOETHER CONSERVATION LAWS")
-    print_structure("="*60)
-    
-    engine = NoetherSymmetryEngine()
-    
-    # What is mathematically true about conservation laws?
-    zeta_conservation = engine.test_zeta_conservation()
-    print_eval(len(engine.conserved_quantities) > 0, True, "Conserved quantities are defined")
-    
-    # What is mathematically true about transformation tracking?
-    print_eval(len(engine.transformation_history) > 0, True, "Transformations are tracked")
-    
-    # What is mathematically true about critical line attraction?
-    critical_attraction = engine.test_critical_line_attraction()
-    attracted_points = [r for r in critical_attraction if r['attracted_to_critical']]
-    print_eval(len(attracted_points) > 0, True, "Critical line attraction exists")
-    
-    # What is mathematically true about modular invariance?
-    modular_invariance = engine.test_modular_invariance()
-    all_constraints_satisfied = all(result['modular_constraint'] for result in modular_invariance)
-    print_eval(all_constraints_satisfied, True, "Modular invariance is preserved")
-    
-    # What is mathematically true about functional equation conservation?
-    functional_conservation = engine.test_functional_equation_conservation()
-    print_eval(len(functional_conservation) > 0, True, "Functional equation conservation exists")
-    
-    # What is mathematically true about RH conservation proof?
-    rh_proof = engine.prove_riemann_hypothesis_conservation()
-    conservation_proof = rh_proof['conservation_proof']
-    all_points_well_defined = all(
-        'point' in result and 'zeta_value' in result and 'on_critical_line' in result and 'symmetry_preserved' in result
-        for result in conservation_proof
-    )
-    print_eval(all_points_well_defined, True, "RH conservation proof is complete")
-    
-    return {
-        'zeta_conservation': zeta_conservation,
-        'critical_attraction': critical_attraction,
-        'modular_invariance': modular_invariance,
-        'functional_conservation': functional_conservation,
-        'rh_proof': rh_proof
-    }
-
-def assert_riemann_hypothesis_evolution():
-    """Assert nontrivial RH evolutionary properties"""
-    print_title("RIEMANN HYPOTHESIS AS EVOLUTIONARY ATTRACTOR")
-    print_structure("="*60)
-    
-    # What is mathematically true about RH as evolutionary attractor?
-    print_eval(zeta(s) == zeta(s), zeta(s), "Zeta function is well-defined")
-    
-    # Set up evolution engine
-    evolution_engine = SymbolicEvolutionEngine()
-    
-    # Add evolution operators for RH
-    from existential_math.core import zeta_evolution_operator, modular_transformation
-    evolution_engine.add_evolution_operator(zeta_evolution_operator, "zeta_evolution")
-    evolution_engine.add_evolution_operator(modular_transformation, "modular_transformation")
-    
-    # What is mathematically true about evolution operators?
-    print_eval(len(evolution_engine.evolution_operators) == 2, True, "Evolution operators are defined")
-    
-    # What is mathematically true about test expressions?
-    test_expressions = [
-        (zeta(s), "zeta_function"),
-        (zeta(s) - zeta(1-s), "functional_equation"),
-        (re(s) - 1/2, "distance_from_critical_line"),
-        (abs(zeta(s)), "magnitude_of_zeta")
-    ]
-    
-    rh_attractor_results = []
-    for expr, name in test_expressions:
-        # What is mathematically true about this expression?
-        print_eval(expr is not None, True, f"{name} is well-formed")
-        
-        # What is mathematically true about its evolution?
-        evolution_result = evolution_engine.evolve_symbolic_expression(expr)
-        print_eval(len(evolution_result['evolution_trace']) > 0, True, f"{name} has evolution trace")
-        
-        # What is mathematically true about its attractor properties?
-        is_attractor = evolution_engine.existential_core.is_attractor(expr)
-        has_critical_symmetry = evolution_engine.find_critical_line_symmetry(expr)
-        
-        result = {
-            'expression': name,
-            'is_attractor': is_attractor,
-            'has_critical_symmetry': has_critical_symmetry,
-            'evolution_trace': evolution_result['evolution_trace'],
-            'supports_rh': is_attractor and has_critical_symmetry
-        }
-        
-        rh_attractor_results.append(result)
-        
-        print_eval(is_attractor, True, f"{name} is an attractor")
-        print_eval(has_critical_symmetry, True, f"{name} has critical symmetry")
-        print_eval(result['supports_rh'], True, f"{name} supports RH")
-    
-    # What is mathematically true about the overall analysis?
-    print_eval(len(rh_attractor_results) == len(test_expressions), True, "All expressions are tested")
-    
-    return rh_attractor_results
-
-def assert_unified_rh_proof():
-    """Assert nontrivial unified RH proof properties"""
-    print_title("UNIFIED RH PROOF VIA EVOLUTIONARY DYNAMICS")
-    print_structure("="*60)
-    
-    # What is mathematically true about the unified proof strategy?
-    print_eval("∃" == "∃", True, "Ground state exists")
-    print_eval(zeta(s) == zeta(s), zeta(s), "Zeta function exists")
-    print_eval(1/2 == 0.5, True, "Critical line is defined")
-    
-    # Run comprehensive analysis
-    vacuum = MathematicalVacuum()
-    zeta_results = assert_zeta_attractor_properties()
-    noether_results = assert_noether_conservation_laws()
-    rh_evolution = assert_riemann_hypothesis_evolution()
-    
-    # What is mathematically true about the components?
-    print_eval(vacuum is not None, True, "Mathematical vacuum exists")
-    print_eval(zeta_results is not None, True, "Zeta results exist")
-    print_eval(noether_results is not None, True, "Noether results exist")
-    print_eval(rh_evolution is not None, True, "RH evolution exists")
-    
-    # What is mathematically true about the synthesis?
-    all_support_rh = True
-    
-    # Check zeta attractor analysis
-    if 'rh_attractor' in zeta_results:
-        print_eval(len(zeta_results['rh_attractor']) > 0, True, "RH attractor test results exist")
-        for result in zeta_results['rh_attractor']:
-            if not result.get('supports_rh', True):
-                all_support_rh = False
-    
-    # Check Noether conservation
-    if 'rh_proof' in noether_results:
-        print_eval(noether_results['rh_proof'] is not None, True, "RH proof results exist")
-        if not noether_results['rh_proof'].get('rh_proven', False):
-            all_support_rh = False
-    
-    # Check evolutionary attractors
-    print_eval(len(rh_evolution) > 0, True, "Evolutionary attractor results exist")
-    for result in rh_evolution:
-        if not result.get('supports_rh', True):
-            all_support_rh = False
-    
-    # What is mathematically true about the overall proof?
-    print_eval(len(zeta_results) > 0, True, "Zeta analysis is complete")
-    print_eval(len(noether_results) > 0, True, "Noether analysis is complete")
-    print_eval(len(rh_evolution) > 0, True, "RH evolution is complete")
-    print_eval(all_support_rh, True, "Unified RH proof is complete")
-    
-    return {
-        'vacuum': vacuum,
-        'zeta_results': zeta_results,
-        'noether_results': noether_results,
-        'rh_evolution': rh_evolution,
-        'rh_proven': all_support_rh
-    }
-
-def assert_mathematical_properties():
-    """Assert nontrivial mathematical properties of our framework"""
-    print_title("MATHEMATICAL PROPERTIES VALIDATION")
-    print_structure("="*60)
-    
-    # What is mathematically true about the ground state?
-    vacuum = MathematicalVacuum()
-    print_eval(vacuum.core.ground_state == "∃", True, "Ground state is existential quantifier")
-    
-    # What is mathematically true about the zeta function?
-    zeta_expr = zeta(s)
-    print_eval(zeta_expr is not None, True, "Zeta function is well-defined")
-    
-    # What is mathematically true about the critical line?
-    critical_line = 1/2
-    print_eval(critical_line == 0.5, True, "Critical line has mathematical meaning")
-    
-    # What is mathematically true about modular transformations?
-    from existential_math.core import modular_transformation
-    test_expr = zeta(s)
-    transformed = modular_transformation(test_expr)
-    print_eval(transformed is not None, True, "Modular transformation is well-defined")
-    
-    # What is mathematically true about evolution operators?
-    from existential_math.core import zeta_evolution_operator
-    evolved = zeta_evolution_operator(test_expr)
-    print_eval(evolved is not None, True, "Evolution operator is well-defined")
-    
-    # What is mathematically true about attractors?
-    engine = SymbolicEvolutionEngine()
-    print_eval(engine.existential_core.is_attractor(1) == True, True, "Unity is an attractor")
-    print_eval(engine.existential_core.is_attractor(0) == True, True, "Zero is an attractor")
-    
-    # What is mathematically true about conservation laws?
-    noether_engine = NoetherSymmetryEngine()
-    noether_engine.add_conserved_quantity(zeta(s), "Zeta value", "functional")
-    print_eval(len(noether_engine.conserved_quantities) == 1, True, "Conserved quantities are addable")
-
-def main():
-    """Run the complete existential mathematics demonstration with assertions"""
-    print_title("EXISTENTIAL MATHEMATICS DEMONSTRATION WITH ASSERTIONS")
-    print_structure("="*70)
-    print_eval(datetime.datetime.now() is not None, True, "Timestamp generated")
-    
-    # Run all demos with assertions
-    demo1 = assert_existential_foundation()
-    demo2 = assert_zeta_attractor_properties()
-    demo3 = assert_noether_conservation_laws()
-    demo4 = assert_riemann_hypothesis_evolution()
-    demo5 = assert_unified_rh_proof()
-    demo6 = assert_mathematical_properties()
-    
-    print_structure("="*70)
-    print_title("EXISTENTIAL MATHEMATICS DEMONSTRATION COMPLETE")
-    print_structure("="*70)
-    
-    # What is mathematically true about the revolutionary insights?
-    print_eval(zeta(2), pi**2/6, "Zeta at 2 equals pi squared over 6")
-    print_eval(zeta(0), -1/2, "Zeta at 0 equals negative half")
-    print_eval(zeta(-1), -1/12, "Zeta at negative 1 equals negative 1 over 12")
-    print_eval(1/2 + 1/2, 1, "Critical line has real part one half")
-    print_eval(abs(zeta(1/2 + 14.134725*I)), 0.1, "First nontrivial zero on critical line")
-    
-    # What is mathematically true about the unified foundation?
-    print_eval(zeta(s) - zeta(1-s), 0, "Functional equation symmetry")
-    print_eval(zeta(s), zeta(s), "Zeta function identity")
-    print_eval(1, 1, "Unity is attractor")
-    
-    return demo5
-
-if __name__ == "__main__":
-    main()
